res fixed a number of compile warnings and soft3d compilation   in optimize mode.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@25249 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/thing/object/lghtmap.cpp
@@ -254,7 +254,7 @@ const char* csLightMap::ReadFromCache (
   csThing* parent = poly->GetParent ();
 
   SetSize (w, h);
-  long lm_size = lwidth * lheight;
+  size_t lm_size = lwidth * lheight;
 
   strcpy (pswanted.header, LMMAGIC);
   if (poly)
